Added edit profile page and functions. Split out form from create user page.
